// ecosystem/lava-sdk/bin/src/codec/subscription/subscription.d.ts
export interface Subscription {
    /** creator pays for the subscription */
    creator: string;
    /** consumer uses the subscription */
    consumer: string;
    /** when the subscription was created */
    block: Long;
    /** index (name) of plan */
    planIndex: string;
    /** when the plan was created */
    planBlock: Long;
    /** total requested duration in months */
    durationTotal: Long;
    /** remaining duration in months */
    durationLeft: Long;
    /** expiry time of current month */
    monthExpiryTime: Long;
    /** when previous month expired */
    prevExpiryBlock: Long;
    /** CU allowance during current month */
    monthCuTotal: Long;
    /** CU remaining during current month */
    monthCuLeft: Long;
    /** CU remaining for previous month */
    prevCuLeft: Long;
}